Update HomeController.java
homeController 위치변경
